Fix things after refactor
