+++ crawler/main.py
@@ -8,14 +8,14 @@ from crawler.constants import (
     COLLECTION_CENTRES,
     COLLECTION_SAMPLES,
     COLLECTION_SOURCE_PLATES,
+    FIELD_BARCODE,
     FIELD_CENTRE_NAME,
     FIELD_LAB_ID,
+    FIELD_LH_SOURCE_PLATE_UUID,
     FIELD_PLATE_BARCODE,
     FIELD_RESULT,
     FIELD_RNA_ID,
     FIELD_ROOT_SAMPLE_ID,
-    FIELD_LH_SOURCE_PLATE_UUID,
-    FIELD_BARCODE,
 )
 from crawler.db import (
     create_mongo_client,
@@ -25,7 +25,7 @@ from crawler.db import (
     samples_collection_accessor,
 )
 from crawler.file_processing import Centre
-from crawler.helpers import get_config
+from crawler.helpers.general_helpers import get_config
 
 logger = logging.getLogger(__name__)
